+++ internal/services/user/service.go
@@ -63,31 +63,31 @@ func (s *service) Get(req GetReq) (resp GetResp, err error) {
 }
 
 func (s *service) Add(req AddReq) (err error) {
-	return base.Callback1(db.GetDB().Create(req.transform()).Error, &req, unilog.Callback[*AddReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(db.GetDB().Create(req.transform()).Error, &req, unilog.Callback[*AddReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) Update(req UpdateReq) (err error) {
-	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(req.transform()).Error, &req, unilog.Callback[*UpdateReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(req.transform()).Error, &req, unilog.Callback[*UpdateReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) UpdatePassword(req UpdatePasswordReq) (err error) {
-	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(&models.User{Password: pkg.MD5(req.Password), UpdateTime: pkg.TimeNowStr()}).Error, &req, unilog.Callback[*UpdatePasswordReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(&models.User{Password: pkg.MD5(req.Password), UpdateTime: pkg.TimeNowStr()}).Error, &req, unilog.Callback[*UpdatePasswordReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) UpdateRole(req UpdateRoleReq) (err error) {
-	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(&models.User{RoleId: req.RoleId, UpdateTime: pkg.TimeNowStr()}).Error, &req, unilog.Callback[*UpdateRoleReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(db.GetDB().Model(&models.User{Id: req.Id}).Updates(&models.User{RoleId: req.RoleId, UpdateTime: pkg.TimeNowStr()}).Error, &req, unilog.Callback[*UpdateRoleReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) Delete(req DeleteReq) (err error) {
-	return base.Callback1(db.GetDB().Delete(&models.User{Id: req.Id}).Error, &req, unilog.Callback[*DeleteReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(db.GetDB().Delete(&models.User{Id: req.Id}).Error, &req, unilog.Callback[*DeleteReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) Enable(req EnableReq) (err error) {
-	return base.Callback1(s.updateState(req.Id, models.UserStateEnable), &req, unilog.Callback[*EnableReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(s.updateState(req.Id, models.UserStateEnable), &req, unilog.Callback[*EnableReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) Disable(req DisableReq) (err error) {
-	return base.Callback1(s.updateState(req.Id, models.UserStateDisable), &req, unilog.Callback[*DisableReq](), base.TransformCallback(&req, req.Callback))
+	return base.Callback1(s.updateState(req.Id, models.UserStateDisable), &req, unilog.Callback[*DisableReq](unilog.Type1Admin()), base.TransformCallback(&req, req.Callback))
 }
 
 func (s *service) updateState(id uint, state byte) (err error) {
@@ -185,7 +185,7 @@ func (s *service) Login(req LoginReq) (resp LoginResp, err error) {
 		return
 	}
 	resp.Token = pkg.MD5(pkg.RandStr(32))
-	_ = base.Callback1(nil, &req, unilog.Callback[*LoginReq]())
+	_ = base.Callback1(nil, &req, unilog.Callback[*LoginReq](unilog.Type1Admin()))
 	_ = base.Callback3(nil, req, user, &resp, req.Callback.Success)
 	return
 }
